Refactor of common functionality to separate files
