+++ Gemfile
@@ -1,4 +1,3 @@
 source 'https://rubygems.org'
 
 gemspec
-gem "stud", :path => "../ruby-stud"
